wip: Fix floating point error in arc support
During the calculations of the planned path for an arc move (say for a spiral Z-Hop), it
is possible that users will encounter a silent failure where the arc move simply does not
occur. This is due to a potential issue arising from floating point precision which
causes the arc support library to calculate the intended path with no arc.

One possible solution, and the solution implemented in this commit, is to simply use
the Decimal class to ensure that floating point precision errors are avoided. This
commit mainly updates existing variables/values to convert them to being Decimal
types. In limited testing so far, this has not caused any additional errors.

Something to be revisited is whether or not the specific style of converting to the
Decimal class is correct for each instance. It likely is fine for the type of
precision used by 3D printers but may not be the most ideal in a vacuum. Also, the
precision used was arbitrarily set to 10 which, again, should be fine for 3D printers
but could also be parameterized or even reduced.